Fixe thread indix invalidation
Continuouse therad stores the indices of the sensors and mappes them to the streams. If a sensor is removed the number of elements in the sensors and processes list decreases. This invalidates the indices stored by the thread run() method. The probem was solved by storing a separate list of sensors with streams. The list gets updated for each thread loop.